disasm: Guess function size
It's to enhance previous commmit by guessing function size instead of
limiting 4096 blindly.

As ELF does not provide function size, it is OK to find next symbol and
then the function size would be kaddr of next symbol minus current
kaddr. And, trim the meaningless tailing int3 and nop insns.

Signed-off-by: Leon Hwang <hffilwlqm@gmail.com>