vmlinux: Search vmlinux and kernel module dbgsym files in the specified directory
In some cases—such as during kernel development—debug symbol (`.ko`) files
are not located in conventional directories. To address this, it's more
reliable to search for `vmlinux` and kernel module debug symbols within
the directory specified by `--kernel-vmlinux`.

For example, when running a self-built kernel along with its debug symbols,
`vmlinux` and related files are typically located under the build output
directory. This PR ensures that such files are found by searching in the
directory provided via `--kernel-vmlinux`.

Signed-off-by: Leon Hwang <leon.hwang@linux.dev>