bpfsnoop: Fix running both entry and exit mode
The issue was introduced by commit dd42c0bdf772 ("bpfsnoop: Run both entry and exit mode").

The event data outputting logic of fexit was broken when run both entry
and exit mode, like `-m 'entry,exit'`, because the data of retval may not
be handled.

To fix it, separate the data size of func args to entry and exit.

Signed-off-by: Leon Hwang <leon.hwang@linux.dev>