fngraph: Fix showing fgraph for the last entry of /proc/kallsyms
It is unable to guess the function insn count of the last entry of
/proc/kallsyms, which is a bpf trampoline symbol possibly.

Then, when the last one is a bpf trampoline symbol, it cannot find its
BTF info as it is not a normal kernel function builtin vmlinux.

Thereafter, as for `--fgraph-proto`, showing its function name is enough.

Signed-off-by: Leon Hwang <leon.hwang@linux.dev>