bpfsnoop: Harden modifying prog insns
When the 'ret' insn is not the last one of a function, there will be
some garbage insns left, which will pollute other function.

In order to find the whole insn range of a function, find the true last
insn index instead, which is the last index of current prog, or the
previous index of the next function's first insn.

Signed-off-by: Leon Hwang <leon.hwang@linux.dev>